requirements.txt
a2a-sdk[http-server]>=0.3.20
uvicorn>=0.28.0
httpx>=0.28.1
